Monitoring only balance of substrate accounts
